Deprecate obsolete methods.
