+++ tests/integration/test_enrollment/data/wazuh_conf.yaml
@@ -25,136 +25,3 @@
                     value: 1
                 - retry_interval:
                     value: 1
-    - tags:
-      - all
-      apply_to_modules:
-      - test_agentd_reconnection
-      sections:
-      - section: client
-        elements:
-            - notify_time:
-                value: 10
-            - time-reconnect:
-                value: 60
-            - auto_restart:
-                value: 'yes'
-            - crypto_method:
-                value: 'aes'
-            - server:
-                elements:
-                - address:
-                    value: SERVER_ADDRESS
-                - port:
-                    value: REMOTED_PORT
-                - protocol:
-                    value: PROTOCOL
-      - section: active-response
-        elements:
-            - disabled:
-                value: 'no'
-      - section: rootcheck
-        elements:
-            - disabled:
-                value: 'no'
-      - section: syscheck
-        elements:
-            - disabled:
-                value: 'no'
-    - tags:
-      - all
-      apply_to_modules:
-      - test_agentd_multi_server
-      sections:
-      - section: client
-        elements:
-            - notify_time:
-                value: 10
-            - time-reconnect:
-                value: 60
-            - auto_restart:
-                value: 'yes'
-            - crypto_method:
-                value: 'aes'
-            - server:
-                elements:
-                - address:
-                    value: SERVER_ADDRESS_1
-                - port:
-                    value: REMOTED_PORT_1
-                - protocol:
-                    value: PROTOCOL
-                - max_retries:
-                    value: 3
-                - retry_interval:
-                    value: 3
-            - server:
-                elements:
-                - address:
-                    value: SERVER_ADDRESS_2
-                - port:
-                    value: REMOTED_PORT_2
-                - protocol:
-                    value: PROTOCOL
-                - max_retries:
-                    value: 3
-                - retry_interval:
-                    value: 3
-            - server:
-                elements:
-                - address:
-                    value: SERVER_ADDRESS_3
-                - port:
-                    value: REMOTED_PORT_3
-                - protocol:
-                    value: PROTOCOL
-                - max_retries:
-                    value: 3
-                - retry_interval:
-                    value: 3
-    - tags:
-      - all
-      apply_to_modules:
-      - test_agentd_parametrized_reconnections
-      sections:
-      - section: client
-        elements:
-            - notify_time:
-                value: 10
-            - time-reconnect:
-                value: 60
-            - auto_restart:
-                value: 'yes'
-            - crypto_method:
-                value: 'aes'
-            - enrollment:
-                elements:
-                - enabled:
-                    value: 'ENROLL'
-            - server:
-                elements:
-                - address:
-                    value: '127.0.0.1'
-                - port:
-                    value: '1514'
-                - protocol:
-                    value: PROTOCOL
-                - max_retries:
-                    value: MAX_RETRIES
-                - retry_interval:
-                    value: RETRY_INTERVAL
-    - tags:
-      - all
-      apply_to_modules:
-      - test_agentd_state_config
-      - test_agentd_state
-      sections:
-      - section: client
-        elements:
-            - server:
-                elements:
-                - address:
-                    value: '127.0.0.1'
-                - port:
-                    value: '1514'
-                - protocol:
-                    value: 'tcp'
