+++ tests/integration/test_enrollment/test_agent_auth_enrollment.py
@@ -4,38 +4,17 @@
 
 import pytest
 import ssl
-import subprocess
-import yaml
-import time
 
 from wazuh_testing.tools.configuration import load_wazuh_configurations
 from wazuh_testing.tools.services import control_service
 from wazuh_testing.tools.monitoring import ManInTheMiddle
-from wazuh_testing.tools.monitoring import QueueMonitor
 from conftest import *
 
 # Marks
-
 pytestmark = [pytest.mark.linux, pytest.mark.win32, pytest.mark.tier(level=0), pytest.mark.agent]
 
-SERVER_ADDRESS = '127.0.0.1'
-
-def load_tests(path):
-    """ Loads a yaml file from a path
-    Returns
-    ----------
-    yaml structure
-    """
-    with open(path) as f:
-        return yaml.safe_load(f)
-
-
-test_data_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data')
-configurations_path = os.path.join(test_data_path, 'wazuh_conf.yaml')
 tests = load_tests(os.path.join(test_data_path, 'wazuh_enrollment_tests.yaml'))
-params = [{'SERVER_ADDRESS': SERVER_ADDRESS, }]
-metadata = [{}]
-configurations = load_wazuh_configurations(configurations_path, __name__, params=params, metadata=metadata)
+configurations = load_wazuh_configurations(configurations_path, __name__)
 
 LAST_MESSAGE = None
 CURRENT_TEST_CASE = {}
@@ -62,36 +41,7 @@ def clear_last_message():
     global LAST_MESSAGE
     LAST_MESSAGE = None
 
-def launch_agent_auth(configuration):
-    parse_configuration_string(configuration)
-    parser = AgentAuthParser(server_address=SERVER_ADDRESS, BINARY_PATH=AGENT_AUTH_BINARY_PATH,
-                             sudo=True if platform.system() == 'Linux' else False)
-    if configuration.get('agent_name'):
-        parser.add_agent_name(configuration.get("agent_name"))
-    if configuration.get('agent_address'):
-        parser.add_agent_adress(configuration.get("agent_address"))
-    if configuration.get('auto_method') == 'yes':
-        parser.add_auto_negotiation()
-    if configuration.get('ssl_cipher'):
-        parser.add_ciphers(configuration.get('ssl_cipher'))
-    if configuration.get('server_ca_path'):
-        parser.add_manager_ca(configuration.get('server_ca_path'))
-    if configuration.get('agent_key_path'):
-        parser.add_agent_certificates(configuration.get('agent_key_path'), configuration.get('agent_certificate_path'))
-    if configuration.get('use_source_ip'):
-        parser.use_source_ip()
-    if configuration.get('password'):
-        parser.add_password(configuration['password']['value'], isFile=(configuration['password']['type'] == 'file'),
-                            path=AUTHDPASS_PATH)
-    else:
-        parser.add_password(None, isFile=True, path=AUTHDPASS_PATH)
-    if configuration.get('groups'):
-        parser.add_groups(configuration.get('groups'))
-
-    out = subprocess.Popen(parser.get_command(), stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    out.communicate()
-
-socket_listener = ManInTheMiddle(address=(SERVER_ADDRESS, 1515), family='AF_INET',
+socket_listener = ManInTheMiddle(address=(DEFAULT_VALUES['manager_address'], DEFAULT_VALUES['port']), family='AF_INET',
                                               connection_protocol='SSL', func=receiver_callback)
 
 # fixtures
@@ -128,7 +78,7 @@ def set_keys(test_case):
             for key in keys:
                 keys_file.writelines(key)
             keys_file.close()
-    except IOError as exception:
+    except IOError:
         raise
 
 @pytest.mark.parametrize('test_case', [case for case in tests])
